refactor(molsetup): stage 1 — split into a package
Convert meeko/molsetup.py (2144 LOC, one file) into the meeko/molsetup/
package with one module per concern:

- atom.py             Atom dataclass + DEFAULT_* atom constants
- bond.py             Bond dataclass + DEFAULT_BOND_* constants
- ring.py             Ring, RingClosureInfo + DEFAULT_RING_CLOSURE_* constants
- restraint.py        Restraint dataclass
- uniq_atom_params.py UniqAtomParams row/column table
- setup.py            MoleculeSetup, MoleculeSetupExternalToolkit,
                      RDKitMoleculeSetup (orchestration classes)
- __init__.py         re-exports for backward compat

No behavior change. External imports `from meeko.molsetup import Atom, ...`
continue to work unchanged. All 92 tests pass (4 skipped, same as baseline).

This is preparatory: subsequent stages will introduce a typed FlexibilityModel
dataclass, extract RDKit interop into an rdkit_adapter module (removing the
MoleculeSetupExternalToolkit ABC), and split MoleculeSetup's chemistry-aware
methods into editing.py / graph.py / io.py.
